+++ PyReconstruct/modules/gui/dialog/__init__.py
@@ -20,4 +20,4 @@ from .shortcuts import ShortcutsDialog
 from .backup_comment import BackupCommentDialog
 from .table_columns import TableColumnsDialog
 from .import_series import ImportSeriesDialog
-from .malformed_contours import MalformedContoursDialog
+from .malformed_contours import MalformedContoursDialog, PixelDustDialog